token mananger with client message proto

+++ internal/define/token_define.go
@@ -12,5 +12,5 @@ const (
 type TokenEntry struct {
 	ID      int32  `json:"_id"`
 	Name    string `json:"name"`
-	MaxHold int32  `json:"max_hold"`
+	MaxHold int64  `json:"max_hold"`
 }
